Move feature-flags location
