+++ src/app/products/layout.tsx
@@ -1,7 +1,7 @@
 import { LeverageBanner } from '@/components/banners/leverage-banner'
 import { Footer } from '@/components/footer'
 import { Header } from '@/components/header'
-import { isLeverageSuiteEnabled } from '@/lib/utils/feature-flags'
+import { isLeverageSuiteEnabled } from '@/feature-flags'
 
 import { ProvidersLite } from '../providers-lite'
